requirements.txt
# xBasis API Dependencies

# Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Auth
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0

# HTTP Client
httpx>=0.26.0

# Redis
redis>=5.0.0

# AI
anthropic>=0.18.0
openai>=1.10.0

# Utils
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Monitoring
sentry-sdk[fastapi]>=1.40.0

# Email
resend>=0.7.0

# Docker SDK
docker>=7.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.1.0
ruff>=0.1.0